requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
pandas>=2.0
openpyxl>=3.0
python-dotenv>=1.0
Werkzeug>=2.0
xlrd==2.0.1